Introduce NetworkError::from_hyper_error and from_ssl_error
