+++ components/net_traits/Cargo.toml
@@ -22,6 +22,7 @@ lazy_static = "0.2"
 log = "0.3.5"
 msg = {path = "../msg"}
 num-traits = "0.1.32"
+openssl = "0.7.6"
 parse-hosts = "0.3.0"
 serde = "0.9"
 serde_derive = "0.9"
